Render the parking-notification map with the self-hosted basemap
#3929 moved the registration map off Mapbox GL onto MapLibre + our
Protomaps PMTiles basemap; the parking-notification pin map was still
loading Mapbox GL tiles. Point it at MAPS_STYLE_URL too, and extract the
MapLibre/PMTiles loader both controllers now share into utils/maplibre.
Mapbox stays only for reverse-geocoding the pin.
