+++ app/components/registrations/show/org_top_actions_parking_notification_form/component.html.erb
@@ -7,6 +7,7 @@
   data-registrations--show--parking-notification-impound-heading-value="<%= translation(".impound_this_bike_type", bike_type: @bike.type) %>"
   data-registrations--show--parking-notification-default-kind-value="<%= new_parking_notification.kind %>"
   data-registrations--show--parking-notification-mapbox-key-value="<%= ENV["MAPBOX_MAPPING"] %>"
+  data-registrations--show--parking-notification-style-url-value="<%= MAPS_STYLE_URL %>"
   data-registrations--show--parking-notification-org-latitude-value="<%= org_map_coordinates[:latitude] %>"
   data-registrations--show--parking-notification-org-longitude-value="<%= org_map_coordinates[:longitude] %>"
   data-action="registrations--show--action-panels:shown->registrations--show--parking-notification#applyMode"
